requirements.txt
opencv-python>=4.8.0,<5.0
pytesseract>=0.3.10
Pillow>=10.0.0
pyautogui>=0.9.53
numpy>=1.24.0,<2.0
pynput>=1.7.6
keyboard>=0.13.5
matplotlib>=3.7.0
pandas>=2.0.0